requirements.txt
black==25.1.0
blinker==1.9.0
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
directory_tree==1.0.0
distlib==0.3.9
filelock==3.18.0
flake8==7.1.2
Flask==3.1.0
fpdf==1.7.2
greenlet==3.1.1
identify==2.6.9
idna==3.10
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pre_commit==4.2.0
pycodestyle==2.12.1
pyflakes==3.2.0
python-dotenv==1.1.0
PyYAML==6.0.2
qrcode==8.0
requests==2.32.3
SQLAlchemy==2.0.39
typing_extensions==4.13.0
urllib3==2.3.0
virtualenv==20.29.3
Werkzeug==3.1.3
